Add support for POST request handling (#26)
This PR adds basic (/experimental) support for processing POST requests, using the Django CSRF pattern as its model. It adds a template tag and a context processor to enable a request token that comes in on a GET request (the standard 'magic link' method) to be passed into a form POST, and then stripped back out, so that it can be used. This implementation requires no change to the existing GET handling.