Update default max uses for LOGIN_MODE_SESSION tokens (#40)
Session tokens with a default max_use of 1 are problematic as email clients / services now often ping URLs in emails in order to cache images / detect phishing sites. This results in session tokens expiring before the user has had a chance to click on the link.

The new default is 10 uses, 10 minutes.