Add management command to truncate logs (#54)
This PR adds a new management command called truncate_request_token_log which can be run on a schedule to ensure that logs don't grow excessively. It supports a log cap defined by date (number of days) or raw count ( number of records), or both.

In addition, it adds a new setting, REQUEST_TOKEN_DISABLE_LOGS which can be used to completely disable the capture of log records. Not recommended, but may be useful in testing.